Move RemovedComponents as well
